# docs/swagger.yaml
basePath: /
definitions:
  handler.getAllCategoriesResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
    type: object
  handler.getAllMovementsResponse:
    properties:
      movements:
        items:
          $ref: '#/definitions/models.Movement'
        type: array
      wallet:
        $ref: '#/definitions/models.Wallet'
    type: object
  handler.getAllWalletsResponse:
    properties:
      wallets:
        items:
          $ref: '#/definitions/models.Wallet'
        type: array
    type: object
  handler.getMovementByIdResponse:
    properties:
      movement:
        $ref: '#/definitions/models.Movement'
      wallet:
        $ref: '#/definitions/models.Wallet'
    type: object
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      icon:
        example: "\U0001F6D2"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Groceries
        type: string
      type:
        description: '"income" or "expense"'
        example: expense
        type: string
      updated_at:
        type: string
      usage_count:
        example: 5
        type: integer
      user_id:
        example: 10
        type: integer
    required:
    - name
    type: object
  models.CreateCategoryInput:
    properties:
      icon:
        example: "\U0001F4B0"
        type: string
      name:
        example: Salary
        type: string
      type:
        enum:
        - income
        - expense
        example: income
        type: string
    required:
    - name
    - type
    type: object
  models.CreateMovementInput:
    properties:
      amount:
        example: 150.5
        type: number
      category:
        example: 1
        type: integer
      date:
        example: "2026-01-27T12:00:00Z"
        type: string
      description:
        example: Grocery shopping
        type: string
      type:
        enum:
        - income
        - expense
        - initial
        example: expense
        type: string
    required:
    - amount
    - category
    - date
    - type
    type: object
  models.CreateWalletInput:
    properties:
      balance:
        example: 1000
        minimum: 0
        type: number
      currency:
        example: USD
        type: string
      name:
        example: Salary Card
        type: string
    required:
    - currency
    - name
    type: object
  models.LogoutInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RefreshInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RegisterInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.SignInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.Movement:
    properties:
      amount:
        type: integer
      category_id:
        type: integer
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        description: '"income" или "expense" или "initial"(только при создании кошелька
          с первоначальным балансом)'
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      wallet_id:
        type: integer
    type: object
  models.UpdateMovementInput:
    properties:
      amount:
        example: 200
        type: number
      category_id:
        example: 3
        type: integer
      date:
        example: "2026-01-28T15:00:00Z"
        type: string
      description:
        example: Updated description
        type: string
      type:
        example: income
        type: string
    type: object
  models.Wallet:
    properties:
      balance:
        example: 15000
        type: integer
      created_at:
        type: string
      currency:
        example: USD
        type: string
      display_id:
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Main Wallet
        type: string
      updated_at:
        type: string
      user_id:
        example: 10
        type: integer
    type: object
info:
  contact: {}
  description: Personal finance tracker API. JWT + Postgres + Docker.
  title: Finance Tracker API
  version: "1.0"
paths:
  /api/categories:
    get:
      description: Получить все категории пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllCategoriesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Список категорий
      tags:
      - categories
  /api/categories/:
    post:
      consumes:
      - application/json
      description: Добавить новую категорию
      parameters:
      - description: Name + Type + Icon
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateCategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: integer
            type: object
      security:
      - Bearer: []
      summary: Создать категорию
      tags:
      - categories
  /api/categories/{id}:
    delete:
      description: Удалить категорию
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Удалить категорию
      tags:
      - categories
    get:
      description: Детали конкретной категории
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Получить категорию по ID
      tags:
      - categories
  /api/wallets/:
    get:
      description: Получить все кошельки пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllWalletsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Список кошельков
      tags:
      - wallets
    post:
      consumes:
      - application/json
      description: Добавить новый кошелёк
      parameters:
      - description: Name + Currency
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateWalletInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: integer
            type: object
      security:
      - Bearer: []
      summary: Создать кошелёк
      tags:
      - wallets
  /api/wallets/{id}:
    delete:
      description: Удалить кошелёк и все транзакции
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Удалить кошелёк
      tags:
      - wallets
    get:
      description: Детали конкретного кошелька
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Получить кошелёк по ID
      tags:
      - wallets
  /api/wallets/{wallet_id}/movements/:
    get:
      description: Получить все операции по кошельку
      parameters:
      - description: Wallet ID
        in: path
        name: wallet_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllMovementsResponse'
      security:
      - Bearer: []
      summary: Список транзакций кошелька
      tags:
      - movements
    post:
      consumes:
      - application/json
      description: Пополнение (+) или списание (-) с кошелька
      parameters:
      - description: Wallet ID
        in: path
        name: wallet_id
        required: true
        type: integer
      - description: Сумма + Тип
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateMovementInput'
      produces:
      - application/json
      responses:
        "200":
          description: Movement ID
          schema:
            additionalProperties:
              type: integer
            type: object
      security:
      - Bearer: []
      summary: Создать транзакцию
      tags:
      - movements
  /api/wallets/{wallet_id}/movements/{trId}:
    delete:
      parameters:
      - description: Wallet ID
        in: path
        name: wallet_id
        required: true
        type: integer
      - description: Movement ID
        in: path
        name: trId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
      security:
      - Bearer: []
      summary: Удалить транзакцию
      tags:
      - movements
    get:
      parameters:
      - description: Wallet ID
        in: path
        name: wallet_id
        required: true
        type: integer
      - description: Movement ID
        in: path
        name: trId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getMovementByIdResponse'
      security:
      - Bearer: []
      summary: Транзакция по ID
      tags:
      - movements
    put:
      consumes:
      - application/json
      parameters:
      - description: Wallet ID
        in: path
        name: wallet_id
        required: true
        type: integer
      - description: Movement ID
        in: path
        name: trId
        required: true
        type: integer
      - description: Changes
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMovementInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
      security:
      - Bearer: []
      summary: Обновить транзакцию
      tags:
      - movements
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация и получение JWT токенов
      parameters:
      - description: Credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Отозвать конкретную refresh сессию пользователя (logout here)
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LogoutInput'
      produces:
      - application/json
      responses:
        "200":
          description: Session deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Not Authorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed deleting session
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выйти со конкретного устройства
      tags:
      - auth
  /auth/logout-all:
    post:
      consumes:
      - application/json
      description: Отозвать все активные refresh сессии пользователя (logout everywhere)
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LogoutInput'
      produces:
      - application/json
      responses:
        "200":
          description: All sessions deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Not Authorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed deleting sessions
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выйти со всех устройств
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access по refresh токену
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить токены
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создать аккаунт
      parameters:
      - description: Email + Password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: User ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"